+++ toshi_hazard_store/config.py
@@ -22,3 +22,9 @@ LOCAL_CACHE_FOLDER = os.getenv('NZSHM22_HAZARD_STORE_LOCAL_CACHE')
 
 SQLITE_ADAPTER_FOLDER = os.getenv('THS_SQLITE_FOLDER')
 USE_SQLITE_ADAPTER = boolean_env('THS_USE_SQLITE_ADAPTER')
+
+
+## SPECIAL SETTINGS FOR MIGRATOIN
+SOURCE_REGION = os.getenv('NZSHM22_HAZARD_STORE_MIGRATE_SOURCE_REGION')
+SOURCE_DEPLOYMENT_STAGE = os.getenv('NZSHM22_HAZARD_STORE_SOURCE_STAGE')
+TARGET_REGION = os.getenv('NZSHM22_HAZARD_STORE_MIGRATE_TARGET_REGION')
